% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_solutions_table.R
\name{create_clean_solutions_table_ui}
\alias{create_clean_solutions_table_ui}
\title{Create Clean Solutions Table UI}
\usage{
create_clean_solutions_table_ui(solutions_data, workflow_info)
}
\description{
Create Clean Solutions Table UI
}
