% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_solutions_table.R
\name{create_header_row1}
\alias{create_header_row1}
\title{Create Header Row 1 (Main Categories)}
\usage{
create_header_row1(optimal_col_name, has_cost = NULL)
}
\description{
Create Header Row 1 (Main Categories)
}
