% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_solutions_table.R
\name{extract_clean_solution_data}
\alias{extract_clean_solution_data}
\title{Extract Clean Solution Data}
\usage{
extract_clean_solution_data(results)
}
\arguments{
\item{results}{Complete analysis results object}
}
\value{
List with structured solution data
}
\description{
Extract Clean Solution Data
}
