% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_solutions_table.R
\name{extract_effect_sizes_clean}
\alias{extract_effect_sizes_clean}
\title{Extract Effect Sizes (Clean)}
\usage{
extract_effect_sizes_clean(optimal, user_config, minimizing_param)
}
\description{
Extract Effect Sizes (Clean)
}
