% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_solutions_table.R
\name{extract_experimental_parameters}
\alias{extract_experimental_parameters}
\title{Extract Experimental Parameters}
\usage{
extract_experimental_parameters(
  optimal,
  user_config,
  minimizing_param,
  workflow_info = NULL
)
}
\description{
Extract Experimental Parameters
}
