% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_solutions_table.R
\name{extract_optimal_parameter_value}
\alias{extract_optimal_parameter_value}
\title{Extract Optimal Parameter Value}
\usage{
extract_optimal_parameter_value(optimal, minimizing_param)
}
\description{
Extract Optimal Parameter Value
}
