% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_solutions_table.R
\name{get_param_value}
\alias{get_param_value}
\title{Get Parameter Value (Optimal vs User Config)}
\usage{
get_param_value(
  param_name,
  optimal,
  user_config,
  minimizing_param,
  workflow_info = NULL
)
}
\description{
Get Parameter Value (Optimal vs User Config)
}
