% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_TPM_fc_minimization.R
\name{TPM-fc-minimization}
\alias{TPM-fc-minimization}
\title{TPM and Fold Change Minimization Analysis Functions}
\description{
Helper functions for power+cost TPM/FC minimization workflows (Workflows 6-7)
using cost_power_computation + find_optimal_cost_design integration with cost constraints.
}
