% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_univariate_optimization.R
\name{univariate-optimization}
\alias{univariate-optimization}
\title{Univariate Optimization Functions}
\description{
Functions for univariate optimization workflows (workflows 1-4, 6-9)
that optimize a single parameter while keeping others fixed.
}
