% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analytics.R
\name{use_google_analytics}
\alias{use_google_analytics}
\title{Configure Google Analytics
This function must be called from a Shiny app's UI.
You can call use_google_analytics() from anywhere inside the UI, as long as the final app UI
(HTML code) contains the result of use_google_analytics().}
\usage{
use_google_analytics(ga_id)
}
\arguments{
\item{ga_id}{The Google Analytics Tag ID, something like "G-ABCD12E34F".}
}
\description{
Configure Google Analytics
This function must be called from a Shiny app's UI.
You can call use_google_analytics() from anywhere inside the UI, as long as the final app UI
(HTML code) contains the result of use_google_analytics().
}
